Split off LoadPrideDatasetCard
