+++ src/components/dataset/LoadDatasetsCard.vue
@@ -24,30 +24,7 @@
             </v-tab-item>
             
             <v-tab-item>
-                <v-card flat>
-                    <v-card-text>
-                        <h3>Load data from the PRIDE archive</h3>
-                        <p>You can easily load data from the <a href="http://www.ebi.ac.uk/pride/" target="_blank">PRIDE</a> data repository. Simply enter an assay id (e.g. 8500) in the field below and click the 'Load PRIDE Dataset' button. The corresponding dataset will then be fetched using the PRIDE API and loaded into the search form on the left.</p>
-                        <v-form ref="prideAssayForm" @submit.prevent>
-                            <v-text-field label="Assay id" placeholder="e.g. 8500" :disabled="prideLoading || pendingStore" v-model="prideAssay" :rules="[value => !!value || 'Please enter a valid PRIDE assay number']" clearable></v-text-field>
-                        </v-form>
-                        <div class="card-actions">
-                            <v-btn v-if="!prideLoading" @click="fetchPrideAssay()">
-                                <v-icon left>mdi-cloud-download</v-icon>
-                                Fetch PRIDE dataset
-                            </v-btn>
-                            <v-progress-linear v-if="prideLoading" v-model="prideProgress"></v-progress-linear>
-                        </div>
-                        <dataset-form ref="prideDatasetForm" v-on:peptide-change="pridePeptides = $event" :peptides="pridePeptides" v-on:name-change="prideName = $event" :name="prideName" v-on:save-change="prideSave = $event" :save="prideSave" :loading="prideLoading || pendingStore"></dataset-form>
-                        <div class="card-actions">
-                            <v-btn :disabled="prideLoading || pendingStore" @click="storePrideDataset()">
-                                <v-icon left>mdi-plus</v-icon>
-                                Add to selected datasets
-                            </v-btn>
-                        </div>
-                        <snackbar :timeout="0" ref="prideSnackbar">Loading dataset... <div class="spinner"></div></snackbar>
-                    </v-card-text>
-                </v-card>
+                <load-pride-dataset-card></load-pride-dataset-card>
             </v-tab-item>
             
             <v-tab-item>
@@ -106,6 +83,7 @@ import Snackbar from "../custom/Snackbar.vue";
 import axios from "axios"
 import CreateDatasetCard from "./CreateDatasetCard.vue";
 import LoadSampleDatasetCard from "./LoadSampleDatasetCard.vue";
+import LoadPrideDatasetCard from "./LoadPrideDatasetCard.vue";
 
 import SampleDataset from "../../logic/data-management/SampleDataset";
 import Tooltip from "../custom/Tooltip.vue";
@@ -119,7 +97,8 @@ import { EventBus } from "../EventBus";
         DatasetForm,
         Tooltip,
         CreateDatasetCard,
-        LoadSampleDatasetCard
+        LoadSampleDatasetCard,
+        LoadPrideDatasetCard
     },
     computed: {
         baseUrl: {
@@ -130,14 +109,6 @@ import { EventBus } from "../EventBus";
     }
 })
 export default class LoadDatasetsCard extends Vue {
-    $refs!: {
-        createdDatasetForm: DatasetForm,
-        prideDatasetForm: DatasetForm,
-        prideSnackbar: Snackbar,
-        // TODO update typings once Vuetify typings available
-        prideAssayForm: any
-    }
-
     @Prop({ required: true })
     private selectedDatasets: Assay[];
     @Prop({ required: true })
@@ -154,58 +125,12 @@ export default class LoadDatasetsCard extends Vue {
     private currentTab: number = 0;
 
     private sampleDatasets: SampleDatasetCollection[] = [];
-    private prideAssay: string = "";
-
-    private createPeptides: string = "";
-    private createName: string = "";
-    private createSave: boolean = true;
-
-    private pridePeptides: string = "";
-    private prideName: string = "";
-    private prideSave: boolean = true;
-    private prideLoading: boolean = false;
-    private prideProgress: number = 0;
 
     private pendingStore: boolean = false;
 
     private loadingSampleDatasets: boolean = true;
     private errorSampleDatasets: boolean = false;
     private selectedSampleDataset = {};
-
-    private fetchPrideAssay(): void {
-        if (this.$refs.prideAssayForm.validate()) {
-            this.prideLoading = true;
-            let datasetManager: DatasetManager = new DatasetManager();
-            let prideNumber: number = parseInt(this.prideAssay);
-
-            this.prideName = "PRIDE assay " + prideNumber.toString();
-
-            // @ts-ignore
-            this.$refs.prideSnackbar.show();
-            datasetManager
-                .loadPrideDataset(prideNumber, (progress) => this.prideProgress = progress * 100)
-                .then((peptides) => {
-                    this.pridePeptides = peptides.join("\n");
-                    this.prideLoading = false;
-                    // @ts-ignore
-                    this.$refs.prideSnackbar.destroy();
-                });
-        }
-    }
-
-    private storePrideDataset() {
-        //@ts-ignore
-        if (this.$refs.prideDatasetForm.isValid()) {
-            this.storeDataset(this.pridePeptides, this.prideName, this.prideSave);
-        }
-    }
-
-    private storeCreatedDataset() {
-        // @ts-ignore
-        if (this.$refs.createdDatasetForm.isValid()) {
-            this.storeDataset(this.createPeptides, this.createName, this.createSave);
-        }
-    }
 
     private selectDataset(dataset: Assay): void {
         EventBus.$emit("select-dataset", dataset);
@@ -214,29 +139,6 @@ export default class LoadDatasetsCard extends Vue {
     private deleteDataset(dataset: Assay): void {
         this.$store.dispatch("deleteDataset", dataset);
         EventBus.$emit("delete-dataset", dataset);
-    }
-
-    private storeDataset(peptides: string, name: string, save: boolean): void {
-        this.pendingStore = true;
-
-        let assay: MetaProteomicsAssay = new MetaProteomicsAssay();            
-        let storageType = save ? StorageType.LocalStorage : StorageType.SessionStorage;
-        let storageWriter: StorageWriter = new StorageWriter();
-
-        assay.setPeptides(peptides.split("\n"));
-        assay.setDate(new Date());
-        assay.setStorageType(save ? StorageType.LocalStorage : StorageType.SessionStorage);
-        assay.setName(name);
-
-        assay.visit(storageWriter).then(
-            () => {
-                this.selectDataset(assay);
-                if (save) {
-                    EventBus.$emit("store-dataset", assay);
-                }
-                this.pendingStore = false;
-            }
-        );
     }
 }
 </script>
