+++ src/components/dataset/LoadSampleDatasetCard.vue
@@ -1,8 +1,8 @@
 <template>
     <v-card flat>
         <v-card-text>
-            <div v-if="loadingSampleDatasets" style="display: flex; justify-content: center;">
-                <v-progress-circular indeterminate></v-progress-circular>
+            <div v-if="loadingSampleDatasets" style="display: flex;">
+                <v-progress-circular indeterminate color="primary"></v-progress-circular>
             </div>
             <div v-else-if="errorSampleDatasets">
                 <v-alert type="error">
